Move endpoints to import bp
