+++ enferno/admin/views.py
@@ -5822,18 +5822,3 @@ def api_bulletin_web_import(validated_data: dict) -> Response:
     )
 
     return jsonify({"batch_id": data_import.batch_id}), 202
-
-
-@admin.get("/api/whisper/models/")
-@roles_required("Admin")
-def api_whisper_models() -> Response:
-    """Returns the list of whisper models."""
-    return jsonify({"models": Constants.WHISPER_MODEL_OPTS})
-
-
-@admin.get("/api/whisper/languages/")
-def api_whisper_languages() -> Response:
-    """Returns the list of whisper languages."""
-    from whisper.tokenizer import TO_LANGUAGE_CODE
-
-    return jsonify({"languages": TO_LANGUAGE_CODE})
