+++ tests/admin/test_whisper_endpoints.py
@@ -13,7 +13,7 @@ get_whisper_models_endpoint_roles = [
 @pytest.mark.parametrize("client_fixture, expected_status", get_whisper_models_endpoint_roles)
 def test_get_whisper_models_endpoint(request, client_fixture, expected_status):
     client_ = request.getfixturevalue(client_fixture)
-    response = client_.get("/admin/api/whisper/models/")
+    response = client_.get("/import/api/whisper/models/")
     assert response.status_code == expected_status
     if expected_status == 200:
         assert response.json == {"models": Constants.WHISPER_MODEL_OPTS}
@@ -23,8 +23,8 @@ def test_get_whisper_models_endpoint(request, client_fixture, expected_status):
 
 get_whisper_languages_endpoint_roles = [
     ("admin_client", 200),
-    ("da_client", 200),
-    ("mod_client", 200),
+    ("da_client", 403),
+    ("mod_client", 403),
     ("anonymous_client", 302),
 ]
 
@@ -32,7 +32,7 @@ get_whisper_languages_endpoint_roles = [
 @pytest.mark.parametrize("client_fixture, expected_status", get_whisper_languages_endpoint_roles)
 def test_get_whisper_languages_endpoint(request, client_fixture, expected_status):
     client_ = request.getfixturevalue(client_fixture)
-    response = client_.get("/admin/api/whisper/languages/")
+    response = client_.get("/import/api/whisper/languages/")
     assert response.status_code == expected_status
     if expected_status == 200:
         assert "languages" in response.json
